+++ project/src/components/card/card.tsx
@@ -1,13 +1,16 @@
-function Card(): JSX.Element {
-  return (
+type CardProps = {
+  src: string;
+}
 
+function Card ({ src }: CardProps): JSX.Element {
+  return (
     <article className="cities__card place-card">
       <div className="place-card__mark">
         <span>Premium</span>
       </div>
       <div className="cities__image-wrapper place-card__image-wrapper">
         <a href="#">
-          <img className="place-card__image" src="img/apartment-01.jpg" width={260} height={200} alt="Place image" />
+          <img className="place-card__image" src={src} width={260} height={200} alt="Place" />
         </a>
       </div>
       <div className="place-card__info">
