+++ project/src/mocks/cadr-mock.js
@@ -1,0 +1,7 @@
+export const cardsSrc = [
+  'img/apartment-01.jpg',
+  'img/room.jpg',
+  'img/apartment-02.jpg',
+  'img/apartment-03.jpg',
+  'img/room.jpg'
+];
